refactor: use better wheel_tag name
