+++ tests/test_builder.py
@@ -9,12 +9,12 @@ from types import SimpleNamespace
 import pytest
 
 from scikit_build_core.builder.builder import Builder
-from scikit_build_core.builder.get_wheel_tag import WheelTag
 from scikit_build_core.builder.macos import get_macosx_deployment_target
 from scikit_build_core.builder.sysconfig import (
     get_python_include_dir,
     get_python_library,
 )
+from scikit_build_core.builder.wheel_tag import WheelTag
 
 
 @pytest.mark.parametrize(
